+++ packages/visual-editor/src/components/pageSections/Breadcrumbs.tsx
@@ -1,3 +1,4 @@
+import { useTranslation } from "react-i18next";
 import {
   Body,
   useTemplateProps,
@@ -7,9 +8,12 @@ import {
   VisibilityWrapper,
 } from "@yext/visual-editor";
 import { ComponentConfig, Fields } from "@measured/puck";
+import { AnalyticsScopeProvider } from "@yext/pages-components";
 
 export type BreadcrumbsSectionProps = {
-  separator?: string;
+  analytics?: {
+    scope?: string;
+  };
   liveVisibility: boolean;
 };
 
@@ -22,6 +26,7 @@ const breadcrumbsSectionFields: Fields<BreadcrumbsSectionProps> = {
     ],
   }),
 };
+
 // getDirectoryParents returns an array of objects. If no dm_directoryParents or children of
 // the directory parent are not the expected objects, returns an empty array.
 const getDirectoryParents = (
@@ -57,8 +62,9 @@ function isValidDirectoryParents(value: any[]): boolean {
 // then displays nothing. In the case of a root DM page, there are
 // no dm_directoryParents but there are dm_directoryChildren so
 // that root entity's name will be in the breadcrumbs.
-export const BreadcrumbsComponent = (props: BreadcrumbsSectionProps) => {
-  const { separator = "/" } = props;
+export const BreadcrumbsComponent = () => {
+  const { t } = useTranslation();
+  const separator = "/";
   const { document, relativePrefixToRoot } = useTemplateProps<any>();
   let breadcrumbs = getDirectoryParents(document);
   if (breadcrumbs?.length > 0 || document.dm_directoryChildren) {
@@ -73,7 +79,11 @@ export const BreadcrumbsComponent = (props: BreadcrumbsSectionProps) => {
   }
 
   return (
-    <PageSection as={"nav"} verticalPadding="sm" aria-label="Breadcrumb">
+    <PageSection
+      as={"nav"}
+      verticalPadding="sm"
+      aria-label={t("breadcrumb", "Breadcrumb")}
+    >
       <ol className="flex flex-wrap">
         {breadcrumbs.map(({ name, slug }, idx) => {
           const isLast = idx === breadcrumbs.length - 1;
@@ -83,6 +93,7 @@ export const BreadcrumbsComponent = (props: BreadcrumbsSectionProps) => {
           return (
             <li key={idx} className="flex items-center">
               <MaybeLink
+                eventName={`link${idx}`}
                 href={isLast ? "" : href}
                 // Force body-sm and link-fontFamily for all breadcrumbs
                 className="text-body-sm-fontSize font-link-fontFamily"
@@ -103,17 +114,22 @@ export const BreadcrumbsSection: ComponentConfig<BreadcrumbsSectionProps> = {
   label: "Breadcrumbs",
   fields: breadcrumbsSectionFields,
   defaultProps: {
+    analytics: {
+      scope: "breadcrumbs",
+    },
     liveVisibility: true,
   },
   render: (props) => {
     return (
-      <VisibilityWrapper
-        liveVisibility={props.liveVisibility}
-        isEditing={props.puck.isEditing}
-        iconSize="md"
-      >
-        <BreadcrumbsComponent {...props} />
-      </VisibilityWrapper>
+      <AnalyticsScopeProvider name={props?.analytics?.scope ?? "breadcrumbs"}>
+        <VisibilityWrapper
+          liveVisibility={props.liveVisibility}
+          isEditing={props.puck.isEditing}
+          iconSize="md"
+        >
+          <BreadcrumbsComponent />
+        </VisibilityWrapper>
+      </AnalyticsScopeProvider>
     );
   },
 };