+++ packages/visual-editor/src/components/pageSections/CoreInfoSection.tsx
@@ -1,11 +1,12 @@
+import { useTranslation } from "react-i18next";
 import * as React from "react";
 import { ComponentConfig, Fields } from "@measured/puck";
 import {
   Address,
   AddressType,
+  AnalyticsScopeProvider,
   DayOfWeekNames,
   getDirections,
-  HoursTable,
   HoursType,
 } from "@yext/pages-components";
 import { FaRegEnvelope } from "react-icons/fa";
@@ -25,6 +26,7 @@ import {
   Background,
   YextField,
   VisibilityWrapper,
+  HoursTableAtom,
 } from "@yext/visual-editor";
 
 export interface CoreInfoSectionProps {
@@ -51,13 +53,16 @@ export interface CoreInfoSectionProps {
       showGetDirectionsLink: boolean;
       phoneFormat: "domestic" | "international";
       includePhoneHyperlink: boolean;
-      emailsListLength: number;
+      emailsListLength?: number;
     };
     hours: {
       startOfWeek: keyof DayOfWeekNames | "today";
       collapseDays: boolean;
       showAdditionalHoursText: boolean;
     };
+  };
+  analytics?: {
+    scope?: string;
   };
   liveVisibility: boolean;
 }
@@ -174,11 +179,6 @@ const coreInfoSectionFields: Fields<CoreInfoSectionProps> = {
               { label: "No", value: false },
             ],
           }),
-          emailsListLength: YextField("List Length", {
-            type: "number",
-            min: 0,
-            max: 3,
-          }),
         },
       }),
       hours: YextField("Hours Column", {
@@ -217,6 +217,7 @@ const coreInfoSectionFields: Fields<CoreInfoSectionProps> = {
 };
 
 const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
+  const { t } = useTranslation();
   const document = useDocument();
   const addressHeadingText = resolveYextEntityField<string>(
     document,
@@ -264,12 +265,15 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
           : "md:[&>section]:w-1/3"
       }`}
       background={styles.backgroundColor}
-      aria-label="Core Info Section"
+      aria-label={t("coreInfoSection", "Core Info Section")}
     >
-      <section aria-label="Information Section" className="flex flex-col gap-4">
+      <section
+        aria-label={t("informationSection", "Information Section")}
+        className="flex flex-col gap-4"
+      >
         {addressHeadingText && (
           <EntityField
-            displayName="Heading Text"
+            displayName={t("headingText", "Heading Text")}
             fieldId={data.info.headingText.field}
             constantValueEnabled={data.info.headingText.constantValueEnabled}
           >
@@ -279,7 +283,7 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
         <div className="flex flex-col gap-2 text-body-fontSize font-body-fontWeight font-body-fontFamily">
           {resolvedAddress && (
             <EntityField
-              displayName="Address"
+              displayName={t("address", "Address")}
               fieldId={data.info.address.field}
               constantValueEnabled={data.info.address.constantValueEnabled}
             >
@@ -295,6 +299,7 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
           )}
           {coordinates && styles.info.showGetDirectionsLink && (
             <CTA
+              eventName={`getDirections`}
               className="font-bold"
               link={coordinates}
               label="Get Directions"
@@ -306,7 +311,8 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
         </div>
         {data.info.phoneNumbers && (
           <ul className="flex flex-col gap-4">
-            {data.info.phoneNumbers.map((item) => {
+            {data.info.phoneNumbers.map((item, idx) => {
+              const { t } = useTranslation();
               const resolvedNumber = resolveYextEntityField<string>(
                 document,
                 item.number
@@ -318,13 +324,14 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
               return (
                 <li key={item.label} className="flex gap-2 items-center">
                   <EntityField
-                    displayName="Phone Number"
+                    displayName={t("phoneNumber", "Phone Number")}
                     fieldId={item.number.field}
                     constantValueEnabled={item.number.constantValueEnabled}
                   >
                     <div className={"flex items-center gap-3"}>
                       <div className="flex gap-2 items-center">
                         <PhoneAtom
+                          eventName={`phone${idx}`}
                           backgroundColor={backgroundColors.background2.value}
                           label={item.label}
                           phoneNumber={resolvedNumber}
@@ -342,7 +349,7 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
         )}
         {resolvedEmails && (
           <EntityField
-            displayName="Email List"
+            displayName={t("emailList", "Email List")}
             fieldId={data.info.emails.field}
             constantValueEnabled={data.info.emails.constantValueEnabled}
           >
@@ -350,7 +357,12 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
               {resolvedEmails
                 .slice(
                   0,
-                  Math.min(resolvedEmails.length, styles.info.emailsListLength)
+                  data.info.emails.constantValueEnabled
+                    ? resolvedEmails.length
+                    : Math.min(
+                        resolvedEmails.length,
+                        styles.info.emailsListLength!
+                      )
                 )
                 .map((email, index) => (
                   <li key={index} className={`flex items-center gap-3`}>
@@ -361,6 +373,7 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
                       <FaRegEnvelope className="w-4 h-4" />
                     </Background>
                     <CTA
+                      eventName={`email${index}`}
                       link={email}
                       label={email}
                       linkType="EMAIL"
@@ -373,10 +386,13 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
         )}
       </section>
       {resolvedHours && (
-        <section aria-label="Hours Section" className="flex flex-col gap-4">
+        <section
+          aria-label={t("hoursSection", "Hours Section")}
+          className="flex flex-col gap-4"
+        >
           {hoursHeadingText && (
             <EntityField
-              displayName="Heading Text"
+              displayName={t("headingText", "Heading Text")}
               fieldId={data.hours.headingText.field}
               constantValueEnabled={data.hours.headingText.constantValueEnabled}
             >
@@ -384,19 +400,21 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
             </EntityField>
           )}
           <EntityField
-            displayName="Hours"
+            displayName={t("hours", "Hours")}
             fieldId="hours"
             constantValueEnabled={data.hours.hours.constantValueEnabled}
           >
-            <HoursTable
+            <HoursTableAtom
               hours={resolvedHours}
               startOfWeek={styles.hours.startOfWeek}
               collapseDays={styles.hours.collapseDays}
-              className="text-body-fontSize font-body-fontWeight font-body-fontFamily"
             />
           </EntityField>
           {additionalHoursText && styles.hours.showAdditionalHoursText && (
-            <EntityField displayName="Hours Text" fieldId="additionalHoursText">
+            <EntityField
+              displayName={t("hoursText", "Hours Text")}
+              fieldId="additionalHoursText"
+            >
               <Body className="mt-4 text-body-sm-fontSize">
                 {additionalHoursText}
               </Body>
@@ -405,10 +423,13 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
         </section>
       )}
       {servicesList && servicesList.length > 0 && (
-        <section aria-label="Services Section" className="flex flex-col gap-4">
+        <section
+          aria-label={t("servicesSection", "Services Section")}
+          className="flex flex-col gap-4"
+        >
           {servicesHeadingText && (
             <EntityField
-              displayName="Heading Text"
+              displayName={t("headingText", "Heading Text")}
               fieldId={data.services.headingText.field}
               constantValueEnabled={
                 data.services.headingText.constantValueEnabled
@@ -420,7 +441,7 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
             </EntityField>
           )}
           <EntityField
-            displayName="Text List"
+            displayName={t("textList", "Text List")}
             fieldId={data.services.servicesList.field}
             constantValueEnabled={
               data.services.servicesList.constantValueEnabled
@@ -443,6 +464,35 @@ const CoreInfoSectionWrapper = ({ data, styles }: CoreInfoSectionProps) => {
 export const CoreInfoSection: ComponentConfig<CoreInfoSectionProps> = {
   label: "Core Info Section",
   fields: coreInfoSectionFields,
+  resolveFields: (data, { fields }) => {
+    if (data.props.data.info.emails.constantValueEnabled) {
+      return fields;
+    }
+
+    return {
+      ...fields,
+      styles: {
+        ...fields.styles,
+        objectFields: {
+          // @ts-expect-error ts(2339) objectFields exists
+          ...fields.styles.objectFields,
+          info: {
+            // @ts-expect-error ts(2339) objectFields exists
+            ...fields.styles.objectFields.info,
+            objectFields: {
+              // @ts-expect-error ts(2339) objectFields exists
+              ...fields.styles.objectFields.info.objectFields,
+              emailsListLength: YextField("Emails List Length", {
+                type: "number",
+                min: 0,
+                max: 3,
+              }),
+            },
+          },
+        },
+      },
+    };
+  },
   defaultProps: {
     data: {
       info: {
@@ -512,14 +562,19 @@ export const CoreInfoSection: ComponentConfig<CoreInfoSectionProps> = {
         showAdditionalHoursText: true,
       },
     },
+    analytics: {
+      scope: "coreInfoSection",
+    },
     liveVisibility: true,
   },
   render: (props) => (
-    <VisibilityWrapper
-      liveVisibility={props.liveVisibility}
-      isEditing={props.puck.isEditing}
-    >
-      <CoreInfoSectionWrapper {...props} />
-    </VisibilityWrapper>
+    <AnalyticsScopeProvider name={props.analytics?.scope ?? ""}>
+      <VisibilityWrapper
+        liveVisibility={props.liveVisibility}
+        isEditing={props.puck.isEditing}
+      >
+        <CoreInfoSectionWrapper {...props} />
+      </VisibilityWrapper>
+    </AnalyticsScopeProvider>
   ),
 };