+++ packages/visual-editor/src/utils/VisualEditorProvider.tsx
@@ -5,6 +5,8 @@ import { TailwindConfig } from "./themeResolver.ts";
 import { StreamFields } from "../types/entityFields.ts";
 import { TailwindConfigContext } from "../hooks/useTailwindConfig.tsx";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { I18nextProvider } from "react-i18next";
+import { i18nComponentsInstance } from "./i18nComponents.ts";
 
 type AllOrNothing<T extends Record<string, any>> =
   | T
@@ -23,7 +25,7 @@ type EditorProps = {
 type VisualEditorProviderProps<T> = UniversalProps<T> &
   AllOrNothing<EditorProps>;
 
-const VisualEditorProvider = <T,>({
+const VisualEditorProvider = <T extends Record<string, any>>({
   templateProps,
   entityFields,
   tailwindConfig,
@@ -31,16 +33,22 @@ const VisualEditorProvider = <T,>({
 }: VisualEditorProviderProps<T>) => {
   const queryClient = new QueryClient();
 
+  if (templateProps?.document?.locale) {
+    i18nComponentsInstance.changeLanguage(templateProps.document.locale);
+  }
+
   return (
-    <QueryClientProvider client={queryClient}>
-      <TemplatePropsContext.Provider value={templateProps}>
-        <EntityFieldsContext.Provider value={entityFields}>
-          <TailwindConfigContext.Provider value={tailwindConfig}>
-            {children}
-          </TailwindConfigContext.Provider>
-        </EntityFieldsContext.Provider>
-      </TemplatePropsContext.Provider>
-    </QueryClientProvider>
+    <I18nextProvider i18n={i18nComponentsInstance}>
+      <QueryClientProvider client={queryClient}>
+        <TemplatePropsContext.Provider value={templateProps}>
+          <EntityFieldsContext.Provider value={entityFields}>
+            <TailwindConfigContext.Provider value={tailwindConfig}>
+              {children}
+            </TailwindConfigContext.Provider>
+          </EntityFieldsContext.Provider>
+        </TemplatePropsContext.Provider>
+      </QueryClientProvider>
+    </I18nextProvider>
   );
 };